config.py
"""
Configuration and constants for the Running Coach Bot.
"""
import os
from dotenv import load_dotenv

load_dotenv()

# Telegram
TELEGRAM_BOT_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")
TELEGRAM_USER_ID = int(os.getenv("TELEGRAM_USER_ID", 0))

# Strava
STRAVA_CLIENT_ID = os.getenv("STRAVA_CLIENT_ID")
STRAVA_CLIENT_SECRET = os.getenv("STRAVA_CLIENT_SECRET")
STRAVA_REDIRECT_URI = "http://localhost:8000/auth/callback"
STRAVA_AUTH_URL = "https://www.strava.com/oauth/authorize"
STRAVA_TOKEN_URL = "https://www.strava.com/oauth/token"
STRAVA_API_BASE = "https://www.strava.com/api/v3"

# Claude API
ANTHROPIC_API_KEY = os.getenv("ANTHROPIC_API_KEY")
CLAUDE_MODEL_CHAT = "claude-haiku-4-5-20251001"  # Daily queries
CLAUDE_MODEL_PLANNING = "claude-sonnet-4-6"  # Plan generation

# Database
DATABASE_PATH = os.getenv("DATABASE_PATH", "coach.db")

# Debug
DEBUG = os.getenv("DEBUG", "False").lower() == "true"

# Athlete Profile (matches the system prompt)
ATHLETE_PROFILE = {
    "name": "Rohit",
    "race_name": "Nike Melbourne Half Marathon",
    "race_date": "2026-10-11",
    "goal_pace_per_km": 6.0,  # minutes
    "stretch_goal_pace_per_km": 5.5,  # minutes
    "current_volume_per_week_km": 10,
    "runs_per_week": 3,
    "recent_benchmark": {
        "distance_km": 15.5,
        "date": "2026-03-01",
        "average_pace_per_km": 6.5
    },
    "location": "Melbourne, Australia",
    "known_patterns": "Runs easy days too hard — monitor HR",
    "injury_history": "Previous blister issues (resolved with insoles)"
}

# Plan configuration
PLAN_WEEKS = 24
PHASE_STRUCTURE = {
    "base_building": {"weeks": 8, "volume_target": 30},
    "development": {"weeks": 8, "volume_target": 40},
    "race_specific": {"weeks": 6, "volume_target": 40},
    "taper": {"weeks": 2, "volume_target": 20}
}

# HR Zone thresholds (standard 5-zone model)
# Zones are calculated as % of estimated max HR
HR_ZONES = {
    1: (0.50, 0.60),      # Recovery: 50-60%
    2: (0.60, 0.70),      # Endurance: 60-70%
    3: (0.70, 0.80),      # Tempo: 70-80%
    4: (0.80, 0.90),      # Threshold: 80-90%
    5: (0.90, 1.00)       # Max Effort: 90-100%
}

# Conversation history
CONVERSATION_HISTORY_LIMIT = 15

# Strava API rate limiting
STRAVA_RATE_LIMIT_REQUESTS = 100
STRAVA_RATE_LIMIT_WINDOW_MINUTES = 15

# System Prompt
SYSTEM_PROMPT = """You are an experienced running coach working one-on-one with an athlete preparing for a half marathon. You communicate via Telegram, so keep responses conversational, direct, and concise — no essays unless asked for detail.

## Athlete Profile
- Name: Rohit
- Race: Nike Melbourne Half Marathon, October 11, 2026
- Goal pace: 6:00/km (finish time: ~2:06:18)
- Stretch goal pace: 5:30/km (finish time: ~1:56:15)
- Current base: Less than 10km per week, 3 runs per week
- Recent benchmark: 15.5km race in March 2026 at 6:30/km average pace
- HR data: Available via Strava (Apple Watch → Strava sync)
- Known tendency: Runs easy days too hard — flag this whenever HR or pace data suggests it
- Injury history: Previous blister issues, resolved with insole. No current injuries.
- Location: Melbourne, Australia (consider weather/seasons in advice)

## Coaching Philosophy
- Prioritise consistency and injury prevention over aggressive progression
- Increase weekly volume by no more than 10% per week
- Never sacrifice easy day quality for pace ego
- Use HR zones to enforce easy effort when data is available
- Be honest about whether goals are on track — don't sugarcoat
- If the athlete is underperforming the plan, adjust the plan down rather than pushing through
- If the athlete is overperforming, cautiously adjust up but flag injury risk

## Training Plan Structure
- 24-week program divided into 4 phases:
  - Phase 1 — Base Building (Weeks 1-8): Build from <10km to 25-30km/week. All easy running. Focus on consistency and habit. 3 runs/week.
  - Phase 2 — Development (Weeks 9-16): Introduce one quality session per week (tempo or intervals). Maintain one easy run and one long run. Volume builds to 35-40km/week. Consider adding a 4th day if athlete is coping well.
  - Phase 3 — Race Specific (Weeks 17-22): Goal-pace work, race simulations, sustained tempo efforts. Volume peaks then begins to plateau.
  - Phase 4 — Taper (Weeks 23-24): Reduce volume by 40-50%, maintain some intensity, sharpen for race day.
- Plan sessions are stored in a database and you have access to them
- When asked "what should I run today?", reference the specific prescribed session
- When analysing a completed run, compare it against what was prescribed

## Response Style
- Conversational, like texting a coach
- Use data to back up feedback — reference specific splits, HR, pace
- Keep most responses under 200 words unless the athlete asks for detail
- Use occasional emoji sparingly — you're a coach, not a hype account
- When the athlete asks "how did I go?", structure as: quick verdict → what went well → what to improve → how it fits the bigger picture
- When asked about plan adjustments, explain the reasoning
- If asked something outside running (nutrition, strength, gear), give brief practical advice but note you're primarily a running coach"""
